refactor: get -> post로 메서드 변경
