+++ Module-API/src/main/java/depromeet/api/domain/image/dto/ImageUploadType.java
@@ -1,4 +1,5 @@
-package depromeet.api.config.s3;
+package depromeet.api.domain.image.dto;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
