+++ Module-API/src/main/java/depromeet/api/domain/image/dto/request/IssuePresignedUrlRequest.java
@@ -1,15 +1,15 @@
 package depromeet.api.domain.image.dto.request;
 
 
-import depromeet.api.config.s3.ImageFileExtension;
-import depromeet.api.config.s3.ImageUploadType;
+import depromeet.api.domain.image.dto.ImageFileExtension;
+import depromeet.api.domain.image.dto.ImageUploadType;
 import depromeet.common.annotation.ValidEnum;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class GetPresignedUrlRequest {
+public class IssuePresignedUrlRequest {
 
     @ValidEnum(enumClass = ImageFileExtension.class, message = "Invalid evaluation parameters")
     private ImageFileExtension imageFileExtension;
