+++ Module-API/src/main/java/depromeet/api/domain/image/usecase/IssuePresignedUrlUseCase.java
@@ -1,22 +1,22 @@
 package depromeet.api.domain.image.usecase;
 
 
-import depromeet.api.config.s3.ImageFileExtension;
-import depromeet.api.config.s3.ImageUploadType;
 import depromeet.api.config.s3.S3UploadPresignedUrlService;
-import depromeet.api.domain.image.dto.response.GetPresignedUrlResponse;
+import depromeet.api.domain.image.dto.ImageFileExtension;
+import depromeet.api.domain.image.dto.ImageUploadType;
+import depromeet.api.domain.image.dto.response.IssuePresignedUrlResponse;
 import depromeet.common.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
 
 @UseCase
 @RequiredArgsConstructor
-public class GetPresignedUrlUseCase {
+public class IssuePresignedUrlUseCase {
     private final S3UploadPresignedUrlService uploadPresignedUrlService;
 
-    public GetPresignedUrlResponse execute(
+    public IssuePresignedUrlResponse execute(
             String socialId, ImageFileExtension imageFileExtension, ImageUploadType type) {
 
-        return GetPresignedUrlResponse.from(
+        return IssuePresignedUrlResponse.from(
                 uploadPresignedUrlService.execute(socialId, imageFileExtension, type));
     }
 }
